Adopt SWB as the binary (HDF5) format for Spacewalk files
