+++ js/CNDBParser.js
@@ -2,7 +2,7 @@ import {openH5File} from 'hdf5-indexed-reader'
 import {FileUtils} from 'igv-utils'
 import {SpacewalkGlobals} from './app.js'
 
-class HDF5Parser {
+class CNDBParser {
 
     constructor() {
     }
@@ -11,7 +11,9 @@ class HDF5Parser {
 
         SpacewalkGlobals.url = false === FileUtils.isFilePath(path) ? path : undefined
 
-        const hdf5 = await openH5File(FileUtils.isFilePath(path) ? { file:path } : { url: path })
+        const config = getCNDBReaderConfiguration(path)
+
+        const hdf5 = await openH5File(config)
 
         const { sample, genomeAssembly } = await datasource.initialize(hdf5)
 
@@ -21,4 +23,17 @@ class HDF5Parser {
 
 }
 
-export default HDF5Parser
+function getCNDBReaderConfiguration(path) {
+
+    const config = {}
+
+    if (FileUtils.isFilePath(path)) {
+        config.file = path
+    } else {
+        config.url = path
+    }
+
+    return config
+}
+
+export default CNDBParser
