+++ js/ensembleManager.js
@@ -4,8 +4,10 @@ import { includes } from "./math.js"
 import {hideGlobalSpinner, showGlobalSpinner} from "./utils.js"
 import Parser from './parser.js'
 import Datasource from './datasource.js'
-import HDF5Parser from "./HDF5Parser.js"
-import HDF5Datasource from "./HDF5Datasource.js"
+import SWBParser from "./SWBParser.js"
+import SWBDatasource from "./SWBDatasource.js"
+import CNDBParser from "./CNDBParser.js"
+import CNDBDatasource from "./CNDBDatasource.js"
 
 class EnsembleManager {
 
@@ -15,8 +17,10 @@ class EnsembleManager {
     async loadURL(url, traceKey) {
 
         const extension = FileUtils.getExtension(url)
-        if ('cndb' === extension) {
-            await this.load(url, new HDF5Parser(), new HDF5Datasource(), parseInt(traceKey))
+        if ('swb' === extension) {
+            await this.load(url, new SWBParser(), new SWBDatasource(), parseInt(traceKey))
+        } else if ('cndb' === extension) {
+            await this.load(url, new CNDBParser(), new CNDBDatasource(), parseInt(traceKey))
         } else {
             await this.load(url, new Parser(), new Datasource(), parseInt(traceKey))
         }
